+++ fabric/src/main/java/de/cech12/brickfurnace/platform/FabricPlatformHelper.java
@@ -3,6 +3,7 @@ package de.cech12.brickfurnace.platform;
 import de.cech12.brickfurnace.platform.services.IPlatformHelper;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ItemStackTemplate;
 
 /**
  * The platform service implementation for Fabric.
@@ -26,12 +27,12 @@ public class FabricPlatformHelper implements IPlatformHelper {
 
     @Override
     public boolean hasCraftingRemainingItem(ItemStack stack) {
-        return stack.getRecipeRemainder() != null && !stack.getRecipeRemainder().isEmpty();
+        return stack.getCraftingRemainder() != null && stack.getCraftingRemainder().count() > 0;
     }
 
     @Override
-    public ItemStack getCraftingRemainingItem(ItemStack stack) {
-        return stack.getRecipeRemainder();
+    public ItemStackTemplate getCraftingRemainingItem(ItemStack stack) {
+        return stack.getCraftingRemainder();
     }
 
 }