+++ neoforge/src/main/java/de/cech12/brickfurnace/platform/NeoForgePlatformHelper.java
@@ -2,6 +2,7 @@ package de.cech12.brickfurnace.platform;
 
 import de.cech12.brickfurnace.platform.services.IPlatformHelper;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.ItemStackTemplate;
 import net.neoforged.fml.ModList;
 import net.neoforged.fml.loading.FMLLoader;
 
@@ -27,11 +28,11 @@ public class NeoForgePlatformHelper implements IPlatformHelper {
 
     @Override
     public boolean hasCraftingRemainingItem(ItemStack stack) {
-        return stack.getCraftingRemainder() != null && !stack.getCraftingRemainder().isEmpty();
+        return stack.getCraftingRemainder() != null && stack.getCraftingRemainder().count() > 0;
     }
 
     @Override
-    public ItemStack getCraftingRemainingItem(ItemStack stack) {
+    public ItemStackTemplate getCraftingRemainingItem(ItemStack stack) {
         return stack.getCraftingRemainder();
     }